demonstrate interpolation
